+++ docs/CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 2.0.1
+- Changed a bug in the Observer class that was sometimes comparing the length of a string rather than an array.
+- Fixed a bug in [varpi, node, arg] relations for retrograde orbits.
 - Large speedup to Observer class internal list comprehension by turning a `Quantity` into a `float`
 - Renamed `SpaceRock.writeto` to `SpaceRock.write_to`.
 - Added a compression kwarg to `SpaceRock.write_to`.